@@ -6,6 +6,7 @@ import java.sql.Timestamp;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.time.Instant;
+import java.time.format.DateTimeFormatter;
 import java.util.TimeZone;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -21,9 +22,12 @@ public class ConvertInstantToTimestampUnitTest {
         instant = timestamp.toInstant();
         assertThat(instant.toEpochMilli()).isEqualTo(timestamp.getTime());
 
-        DateFormat df = DateFormat.getDateTimeInstance();
-        df = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SS'Z'");
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'");
+        formatter = formatter.withZone(TimeZone.getTimeZone("UTC").toZoneId());
+
+        DateFormat df = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'");
         df.setTimeZone(TimeZone.getTimeZone("UTC"));
-        assertThat(instant.toString()).isEqualTo(df.format(timestamp).toString());
+
+        assertThat(formatter.format(instant)).isEqualTo(df.format(timestamp));
     }
 }
