@@ -18,6 +18,7 @@ import java.security.NoSuchAlgorithmException;
 import java.time.Duration;
 
 import org.junit.Test;
+import org.junit.jupiter.api.Disabled;
 
 public class HttpRequestUnitTest {
 
@@ -48,7 +49,12 @@ public class HttpRequestUnitTest {
         assertThat(response.version(), equalTo(HttpClient.Version.HTTP_2));
     }
 
-    @Test
+    /*
+    * This test will fail as soon as the given URL returns a HTTP 2 response.
+    * Therefore, let's leave it commented out.
+    * */
+   @Test
+   @Disabled
     public void shouldFallbackToHttp1_1WhenWebsiteDoesNotUseHttp2() throws IOException, InterruptedException, URISyntaxException, NoSuchAlgorithmException {
         HttpRequest request = HttpRequest.newBuilder()
             .uri(new URI("https://postman-echo.com/get"))
