@@ -21,12 +21,8 @@ public class LocalDateTimeUnitTest {
     @Test
     public void givenLocalDateTimeMock_whenNow_thenGetFixedLocalDateTime() {
         Clock clock = Clock.fixed(Instant.parse("2014-12-22T10:15:30.00Z"), ZoneId.of("UTC"));
-        LocalDateTime dateTime = LocalDateTime.now(clock);
-        mockStatic(LocalDateTime.class);
-        when(LocalDateTime.now()).thenReturn(dateTime);
         String dateTimeExpected = "2014-12-22T10:15:30";
-
-        LocalDateTime now = LocalDateTime.now();
+        LocalDateTime now = LocalDateTime.now(clock);
 
         assertThat(now).isEqualTo(dateTimeExpected);
     }
